Recipe_Recommendation_Bot/requirements.txt
beautifulsoup4==4.11.1
numpy==1.21.5
openai==0.18.1
pandas==1.2.5
python-dotenv==0.20.0
recipe_scrapers==13.31.0
requests==2.27.1
spoonacular==3.0
telegram==0.0.1
tensorflow==2.4.1
